+++ attributes/default.rb
@@ -4,7 +4,7 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 
-#     http//www.apache.org/licenses/LICENSE-2.0
+#     http://www.apache.org/licenses/LICENSE-2.0
 
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -26,7 +26,15 @@ default["airflow"]["directories_mode"] = "0775"
 default["airflow"]["config_file_mode"] = "0644"
 default["airflow"]["bin_path"] = node["platform"] == "ubuntu" ? "/usr/local/bin" : "/usr/bin"
 default["airflow"]["run_path"] = "/var/run/airflow"
-default["airflow"]["init_system"] = node["platform"] == "ubuntu" ? "upstart" : "systemd"
+default["airflow"]["is_upstart"] = node["platform"] == "ubuntu" && node["platform_version"].to_f < 15.04
+default["airflow"]["init_system"] = node["airflow"]["is_upstart"] ? "upstart" : "systemd"
+default["airflow"]["env_path"] = node["platform_family"] == "debian" ? "/etc/default/airflow" : "/etc/sysconfig/airflow"
+
+
+# Python config
+default["airflow"]["python_runtime"] = "2"
+default["airflow"]["python_version"] = "2.7"
+default["airflow"]["pip_version"] = true
 
 # Configurations stated below are required for this cookbook and will be written to airflow.cfg, you can add more config by using structure like:
 # default["airflow"]["config"]["CONFIG_SECTION"]["CONFIG_ENTRY"]
@@ -39,4 +47,3 @@ default["airflow"]["config"]["core"]["sql_alchemy_conn"] = "sqlite:///#{node["ai
 default["airflow"]["config"]["core"]["fernet_key"] = "G3jB5--jCQpRYp7hwUtpfQ_S8zLRbRMwX8tr3dehnNU=" # Be sure to change this for production
 # Celery
 default["airflow"]["config"]["celery"]["celeryd_concurrency"] = 16
-