+++ recipes/packages.rb
@@ -1,3 +1,5 @@
+# Copyright 2015 Sergey Bahchissaraitsev
+
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -10,7 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-include_recipe 'poise-python'
+python_runtime node["airflow"]["python_runtime"] do
+  version node["airflow"]["python_version"]
+  provider :system
+  pip_version node["airflow"]["pip_version"]  
+end
 
 # Obtain the current platform name
 platform = node['platform'].to_s